Fix build error: "Liquid Exception: Invalid scheme format"
Some of my post titles include a colon. For example:

> Git tip: How to "merge" specific files from another branch

The `uri_escape` filter was interpreting interpreting these titles as
URLs, and generating an "invalid scheme format" error. In other words,
it thought that I was trying to use "Git tip" as a URL scheme like
"http". `cgi_escape` seems like the more appropriate choice here, and it
fixes the error. From the liquid docs on `cgi_escape`:

> CGI escape a string for use in a URL. Replaces any special characters
> with appropriate %XX replacements.